+++ spi_nand_flash/test_app/pytest_spi_nand_flash.py
@@ -32,6 +32,23 @@ def test_spi_nand_flash(dut: Dut, config: str) -> None:
 @pytest.mark.parametrize(
     'config',
     [
+        'anonymous',
+    ],
+    indirect=True,
+)
+@idf_parametrize('target', ['esp32'], indirect=['target'])
+def test_spi_nand_flash_anonymous(dut: Dut, config: str) -> None:
+    dut.run_all_single_board_cases()
+
+
+@pytest.mark.spi_nand_flash
+@pytest.mark.skipif(
+    not bool(glob.glob(f'{Path(__file__).parent.absolute()}/build*/')),
+    reason="Skip the idf version that not build"
+)
+@pytest.mark.parametrize(
+    'config',
+    [
         'bdl',
     ],
     indirect=True,